+++ examples/stream.py
@@ -70,12 +70,9 @@ if __name__=="__main__":
     # logging.basicConfig(format='[%(levelname)-5s] %(name)s: %(message)s',level=logging.DEBUG)
     # logging.getLogger("htag.tag").setLevel( logging.INFO )
 
-    # instanciate the main component
-    obj=Page()
-
     # and execute it in a pywebview instance
     from htag.runners import *
 
     # here is another runner, in a simple browser (thru ajax calls)
-    BrowserHTTP( obj ).run()
-    # PyWebWiew( obj ).run()
+    BrowserHTTP( Page ).run()
+    # PyWebWiew( Page ).run()