+++ htag/runners/browserstarlettehttp.py
@@ -22,8 +22,8 @@ class BrowserStarletteHTTP:
     """ Simple ASync Web Server (with starlette) with HTTP interactions with htag.
         Open the rendering in a browser tab.
     """
-    def __init__(self,tag:Tag):
-        assert isinstance(tag,Tag)
+    def __init__(self,tagClass:type):
+        assert issubclass(tagClass,Tag)
 
         js = """
 async function interact( o ) {
@@ -33,7 +33,7 @@ async function interact( o ) {
 window.addEventListener('DOMContentLoaded', start );
 """
 
-        self.renderer=HRenderer(tag, js, lambda: os._exit(0))
+        self.renderer=HRenderer(tagClass, js, lambda: os._exit(0))
 
     async def GET(self,request) -> HTMLResponse:
         return HTMLResponse( str(self.renderer) )