+++ htag/runners/browserstarlettews.py
@@ -21,8 +21,8 @@ class BrowserStarletteWS:
     """ Simple ASync Web Server (with starlette) with WebSocket interactions with HTag.
         Open the rendering in a browser tab.
     """
-    def __init__(self,tag:Tag):
-        assert isinstance(tag,Tag)
+    def __init__(self,tagClass:type):
+        assert issubclass(tagClass,Tag)
         js = """
 async function interact( o ) {
     ws.send( JSON.stringify(o) );
@@ -35,7 +35,7 @@ ws.onmessage = function(e) {
 };
 """
 
-        self.renderer=HRenderer(tag, js, lambda: os._exit(0))
+        self.renderer=HRenderer(tagClass, js, lambda: os._exit(0))
 
     def run(self, host="127.0.0.1", port=8000, openBrowser=True ):   # localhost, by default !!
         app = Starlette()